App: Fix fast-forward on mobile devices
Pressing on the logo in mobile devices would cause the mobile browser
to register a right-click on an image which brings out a menu.

As a result, users on mobile devices aren't able to activate the
fast-forward function. To fix this, we add an overlay div attached with
press listeners to bypass the right-click behavior on mobile browsers.